LSArray + RoleManager + Data Cleanup

+++ Morph-2020/lib/Data Structs/GoalData.h
@@ -22,7 +22,7 @@ class GoalData{
             return angle != 65535;
         }
 
-        int calculateCentimeter(){
+        float calculateCentimeter(){
             return 0.00000353067 * pow(MATH_E, 0.0288427 * (distance + 500)) - 25.0921;
         }
